Erroneous teapot state
The "unlinked" logic was incorrect in [teapot state].  This was
masked by a bug in TeapotIsLinked.  On Windows (on my machine,
at least), [info nameofexecutable] returns the path with a lowercase
"c:".  Apparently an uppercase "C:" is canonical.  So TeapotIsLinked was
erroneously saying that the teapot was NOT linked, and [teapot state]
was saying that this was good.  Both fixed.