Kite updates project versions in require blocks
When a project provides a package, any packages that use it in the same
project must "package require" it.  If the "package require" statement
is put in the -kite-require block, the version will now be updated to
the project version automatically.  This ensures that the requiring
package will get the development version rather than some previous
stable version in the teapot.