'kite compile' sets no environment vars
The initial version of 'kite compile' set environment variables for
KITE_OS and similar things, so that they would be available in
Makefiles.  It's better, though, for the Makefiles to use 'kite info' to
retrieve the desired values; then the Makefiles can work from the shell
command line as well as from Kite.  Consequently, it's no longer
necessary for 'kite compile' to set them, and it doesn't.

In addition, 'kite compile''s output was improved.