Added a "tclsh" macro to manpage(5)
The new "tclsh" macro runs a script in an external tclsh which is set up
to make the project's code base available.  The result of the script is
included in the document.  This should allow manpages to execute project
code (thus supporting automatic documentation generation) while still
running it against the project's version of Tcl.