Fixes #28: kite docs formats file target
If the first argument to "kite docs" is a file, it will attempt to
format the file.  It must be a .ehtml file under <root>/docs/.  If it's
in a man* directory, all man pages in that directory are formatted.
Otherwise, the specific file is formatted as a kitedoc(5) file.