Test tool updated to be more flexible
The test tool still takes the same arguments it did before, it just does
some more processing to figure out what is wanted for a test target if
there's only one argument before the options.

Basically, there's this:
* No args: run all tests passing down any options to tcltest
* One arg: if arg is target, all tests in target;  if arg is test
module, test dirs are traversed looking for the target that contains
that test. In either case, all options are passed down
* Two args: it is assumed that first arg is target and second is module.